Verified test3183 against hsql, oracle and postgresql

+++ cpactf/src/test/ddl/org/castor/cpa/test/test3183/postgresql-create.sql
@@ -1,16 +1,16 @@
 CREATE TABLE test3183_poly_doc (
-  id    int           NOT NULL, 
-  name  varchar(20)   NOT NULL
+  id    int              NOT NULL, 
+  name  varchar(20)      NOT NULL
 );
 
 CREATE TABLE test3183_poly_order_doc (
-  id    int           NOT NULL, 
-  price double        NOT NULL
+  id    int              NOT NULL, 
+  price double precision NOT NULL
 );
 
 
 CREATE TABLE test3183_single_order_doc (
-  id    int           NOT NULL, 
-  name  varchar(20)   NOT NULL,
-  price double        NOT NULL
+  id    int              NOT NULL, 
+  name  varchar(20)      NOT NULL,
+  price double precision NOT NULL
 );
